Consolidate trigger UGens
